refactor: add type of pages
