+++ gatsby-theme-oi-wiki/src/pages/404.tsx
@@ -1,4 +1,3 @@
-// Components
 import React from 'react'
 import StyledLayout from '../components/StyledLayout'
